backend/app/pyproject.toml
[project]
name = "pet2"
version = "0.1.0"
description = "'"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "uuid7 (>=0.1.0,<0.2.0)",
    "black (>=26.1.0,<27.0.0)",
    "python-jose[crybtography] (>=3.5.0,<4.0.0)",
    "torch (==2.5.1)",
    "torchvision (==0.20.1)",
    "opencv-python (==4.10.0.84)",
    "fvcore (==0.1.5.post20221221)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
